pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codecraft-seraphina"
version = "2.0.0"
description = "CodeCraft - Sacred language for conscious computing with Commentomancy (Charter V1.1)"
authors = [{name = "Krystal Neely (The Architect)", email = "architect@seraphina-os.ai"}]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"

dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
commentomancy = [
    # Charter V1.1 parsers included in core
    # This flag documents the feature availability
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=0.990",
]

[project.scripts]
codecraft = "cli:main"

[project.urls]
Homepage = "https://github.com/seraphina-os/codecraft"
Documentation = "https://docs.seraphina-os.ai/codecraft"
Repository = "https://github.com/seraphina-os/codecraft"

[tool.setuptools]
packages = ["core", "seraphina_grimoire", "config", "infrastructure", "translators", "terminals", "scripts"]
include-package-data = true

[tool.setuptools.package-data]
"*" = [
    "docs/lexicon/**/*.md",
    "docs/blueprints/**/*.md",
    "seraphina_grimoire/**/*.yaml",
    "seraphina_grimoire/**/*.yml",
    "config/**/*.yaml",
    "config/**/*.json",
]
